+++ src/Message/Response.php
@@ -8,11 +8,13 @@ use Omnipay\Common\Message\RequestInterface;
 /**
  * Class Response
  *
- * @author Bram Devries <bram@madewithlove.be>
+ * @author  Bram Devries <bram@madewithlove.be>
  * @package Omnipay\Paymill\Message
  */
 class Response extends AbstractResponse
 {
+	const FAILED = 'failed';
+
 	/**
 	 * Is the response successful?
 	 *
@@ -20,7 +22,15 @@ class Response extends AbstractResponse
 	 */
 	public function isSuccessful()
 	{
-		return !isset($this->data['error']);
+		$data = $this->getData();
+
+		// Check if the request failed
+		if (isset($data['error'])) {
+			return false;
+		}
+
+		// The request succeeded, but the transaction could still fail
+		return $data['status'] !== self::FAILED;
 	}
 
 	/**
@@ -42,7 +52,14 @@ class Response extends AbstractResponse
 	 */
 	public function getMessage()
 	{
-		return !$this->isSuccessful() ? $this->data['error'] : null;
+		$data    = $this->getData();
+		$message = null;
+
+		if (isset($data['error'])) {
+			return $data['error'];
+		}
+
+		return ErrorMessage::getErrorForCode($data['response_code']);
 	}
 
 	/**